+++ apps/dashboard/src/app/layout.tsx
@@ -26,10 +26,17 @@ export const metadata: Metadata = {
     default: 'Stack Auth Dashboard',
     template: '%s | Stack Auth',
   },
-  description: 'Stack Auth is the fastest way to add authentication to your web app.',
+  description: 'Stack Auth is the open-source Auth0 alternative, and the fastest way to add authentication to your web app.',
   openGraph: {
+    title: 'Stack Auth Dashboard',
+    description: 'Stack Auth is the open-source Auth0 alternative, and the fastest way to add authentication to your web app.',
     images: [`${process.env.NEXT_PUBLIC_STACK_URL_DEPRECATED}/open-graph-image.png`]
-  }
+  },
+  twitter: {
+    title: 'Stack Auth Dashboard',
+    description: 'Stack Auth is the open-source Auth0 alternative, and the fastest way to add authentication to your web app.',
+    images: [`${process.env.NEXT_PUBLIC_STACK_URL_DEPRECATED}/open-graph-image.png`]
+  },
 };
 
 const fontSans = FontSans({