+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stackframe/stack-backend
 
+## 2.5.25
+
+### Patch Changes
+
+- GitLab OAuth provider
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.25
+  - @stackframe/stack-emails@2.5.25
+
 ## 2.5.24
 
 ### Patch Changes