+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackframe/stack-ui
 
+## 2.5.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.25
+
 ## 2.5.24
 
 ### Patch Changes